# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SimpleSchedule service API.",
        "title": "SimpleSchedule API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2811",
    "basePath": "/",
    "paths": {
        "/api/slots": {
            "post": {
                "description": "Create a new schedule slot with time range and day validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Create slot",
                "parameters": [
                    {
                        "description": "Slot creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
        "/api/slots/{id}": {
            "get": {
                "description": "Retrieve a single schedule slot with its associated tag by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Get slot by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SlotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of an existing slot with collision validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Update slot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Slot update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a schedule slot by its unique identifier",
                "tags": [
                    "slots"
                ],
                "summary": "Delete slot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
...
        "/api/weeks/{id}/slots": {
            "get": {
                "description": "Retrieve all schedule slots with tags for a given week",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Get slots by week ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Week ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SlotsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apierror.UserError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "domain.DayOfWeek": {
            "type": "string",
            "enum": [
                "monday",
                "tuesday",
                "wednesday",
                "thursday",
                "friday",
                "saturday",
                "sunday"
            ],
            "x-enum-varnames": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday"
            ]
        },
        "domain.SlotWithTag": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dayOfWeek": {
                    "$ref": "#/definitions/domain.DayOfWeek"
                },
                "endTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "startTime": {
                    "type": "integer"
                },
                "tag": {
                    "$ref": "#/definitions/domain.Tag"
                },
                "tagId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "weekId": {
                    "type": "string"
                }
            }
        },
        "domain.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "hexColor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateSlotRequest": {
            "type": "object",
            "required": [
                "activity",
                "dayOfWeek",
                "weekId"
            ],
            "properties": {
                "activity": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Higher Mathematics"
                },
                "dayOfWeek": {
                    "enum": [
                        "monday",
                        "tuesday",
                        "wednesday",
                        "thursday",
                        "friday",
                        "saturday",
                        "sunday"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.DayOfWeek"
                        }
                    ],
                    "example": "monday"
                },
                "endTime": {
                    "type": "integer",
                    "maximum": 86400,
                    "minimum": 0,
                    "example": 39600
                },
                "startTime": {
                    "type": "integer",
                    "maximum": 86400,
                    "minimum": 0,
                    "example": 36000
                },
                "tagId": {
                    "type": "string",
                    "example": "97ec8552-9562-4d79-ab4c-1ca959663629"
                },
                "weekId": {
                    "type": "string",
                    "example": "d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2"
                }
            }
        },
...
        "handlers.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "97ec8552-9562-4d79-ab4c-1ca959663629"
                }
            }
        },
        "handlers.SlotResponse": {
            "type": "object",
            "properties": {
                "slot": {
                    "$ref": "#/definitions/domain.SlotWithTag"
                }
            }
        },
        "handlers.SlotsResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SlotWithTag"
                    }
                }
            }
        },
...
        "handlers.UpdateSlotRequest": {
            "type": "object",
            "required": [
                "activity",
                "dayOfWeek",
                "weekId"
            ],
            "properties": {
                "activity": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Physics Lecture"
                },
                "dayOfWeek": {
                    "enum": [
                        "monday",
                        "tuesday",
                        "wednesday",
                        "thursday",
                        "friday",
                        "saturday",
                        "sunday"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.DayOfWeek"
                        }
                    ],
                    "example": "monday"
                },
                "endTime": {
                    "type": "integer",
                    "maximum": 86400,
                    "minimum": 0,
                    "example": 39600
                },
                "startTime": {
                    "type": "integer",
                    "maximum": 86400,
                    "minimum": 0,
                    "example": 36000
                },
                "tagId": {
                    "type": "string",
                    "example": "97ec8552-9562-4d79-ab4c-1ca959663629"
                },
                "weekId": {
                    "type": "string",
                    "example": "d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2"
                }
            }
        },
...
    }
}